[allura:tickets:#6208] use background=True when ensuring indexes
Mongo defaults to creating indexes as blocking foreground operations
(presumably for backwards compatibility), but this is dangerous
since it is blocking.  Use background=True when ensuring indexes
so that adding a new index to a model doesn't cause the mongo server
to block up as soon as that new model is in use.

http://docs.mongodb.org/manual/core/indexes/#index-creation-background